# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger документация проекта <<В добрые руки>>",
        "description": "Open API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://in-good-hands.dev.mind4.me",
            "description": " API Server"
        }
    ],
    "paths": {
...
        "/api/admin/all_cities": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Получение списка городов",
                "operationId": "16ab9711c5e716f25eb2ecae09ee1d22",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/create_city": {
            "post": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Создание города",
                "operationId": "fffe2e204519cb84eb8a1de3a7bc8cb3",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "is_active"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Вологда"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/delete_city": {
            "delete": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Удаление города",
                "operationId": "a5886de7f05c666fc81b5b57af86a8f1",
                "parameters": [
                    {
                        "name": "id_city",
                        "in": "query",
                        "required": true,
                        "example": 9
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/change_city": {
            "patch": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Изменение города",
                "operationId": "ad612dd3e417cb72f5ffcee98d50a966",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_city",
                                    "name",
                                    "is_active"
                                ],
                                "properties": {
                                    "id_city": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Вологда"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/all_categories": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Получение списка категорий",
                "operationId": "89e60f0fac5aa194a8f121400fc60ff1",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/create_category": {
            "post": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Создание категории",
                "operationId": "e356341a60b21fa5b4a1dab97c890b30",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "icon",
                                    "name",
                                    "is_active"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Новое название"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "👕"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/delete_category": {
            "delete": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Удаление категории",
                "operationId": "6e2b76e33f859777e365b632e45eda0f",
                "parameters": [
                    {
                        "name": "id_category",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/change_category": {
            "patch": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Изменение категории",
                "operationId": "bdd6ed891cb95d8312d72b83fdbaded3",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_category",
                                    "icon",
                                    "name",
                                    "is_active"
                                ],
                                "properties": {
                                    "id_category": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Новое название"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "👕"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/get_all_users": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Получение списка пользователей",
                "operationId": "cad185918bc539ca4ec882c566e75c0c",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 1
                    },
                    {
                        "name": "id_user",
                        "in": "query",
                        "required": false,
                        "example": 123
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "example": "user@mail.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/admin/get_all_posts": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Получение списка постов ",
                "operationId": "fa4a8bfc3157712a018156ff58b3d8bb",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 1
                    },
                    {
                        "name": "id_user",
                        "in": "query",
                        "required": false,
                        "example": 2
                    },
                    {
                        "name": "id_city",
                        "in": "query",
                        "required": false,
                        "example": 2
                    },
                    {
                        "name": "id_category",
                        "in": "query",
                        "required": false,
                        "example": 2
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "example": "Еда"
                    },
                    {
                        "name": "sort_type",
                        "in": "query",
                        "required": false,
                        "example": "asc"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "example": "date"
                    },
                    {
                        "name": "id_ad",
                        "in": "query",
                        "required": false,
                        "example": 123
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/admin/get_post_statusses": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Получение списка статусов поста",
                "operationId": "f78b08d09344352228b60c8ca87d250e",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://in-good-hands.dev.mind4.me/oauth/authorize",
                        "tokenUrl": "https://in-good-hands.dev.mind4.me/oauth/token",
                        "refreshUrl": "https://in-good-hands.dev.mind4.me/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}